requirements.txt
rapidocr-onnxruntime
opencv-python
Pillow
mss
pyautogui
keyboard
